requirement.txt
pyngrok
discord.py
requests